+++ src/pynamicalsys/__init__.py
@@ -17,6 +17,7 @@
 
 from pynamicalsys.core.discrete_dynamical_systems import DiscreteDynamicalSystem
 from pynamicalsys.core.continuous_dynamical_systems import ContinuousDynamicalSystem
+from pynamicalsys.core.hamiltonian_systems import HamiltonianSystem
 from pynamicalsys.core.basin_metrics import BasinMetrics
 from pynamicalsys.core.plot_styler import PlotStyler
 from pynamicalsys.core.time_series_metrics import TimeSeriesMetrics
@@ -25,6 +26,7 @@ from .__version__ import __version__
 __all__ = [
     "DiscreteDynamicalSystem",
     "ContinuousDynamicalSystem",
+    "HamiltonianSystem",
     "PlotStyler",
     "TimeSeriesMetrics",
     "BasinMetrics",